+++ src/version/plugins/Vaccinations/VaccinationsCreate.js
@@ -1,20 +1,10 @@
 import React from "react";
-import { DisabledInput, TextInput, DateInput, LongTextInput } from "react-admin";
-
-import { withStyles } from '@material-ui/core/styles';
 
 import CreateTemplate from "../../../core/common/ResourseTemplates/CreateTemplate";
-
-const styles = {
-    labelBlock: {
-        '& > div': {
-            marginBottom: "0px !important",
-        },
-    },
-};
+import Inputs from "./fragments/Inputs";
 
 /**
- * This component returns Problems creation form
+ * This component returns Vaccinations creation form
  *
  * @author Bogdan Shcherban <bsc@piogroup.net>
  * @param {shape} classes
@@ -22,14 +12,8 @@ const styles = {
  */
 const VaccinationsCreate = ({ classes, ...rest }) => (
     <CreateTemplate blockTitle="Problems / Issues" {...rest}>
-        <TextInput className={classes.labelBlock} source="vaccinationName" label="Name" />
-        <DateInput className={classes.labelBlock} source="vaccinationDateTime" label="Date and Time" />
-        <TextInput className={classes.labelBlock} source="series" label="Series" />
-        <LongTextInput className={classes.labelBlock} source="comment" label="Comment" />
-        <DisabledInput className={classes.labelBlock} source="author" label="Author" />
-        <DisabledInput className={classes.labelBlock} source="source" label="Source" />
-        <DisabledInput className={classes.labelBlock} source="date" label="Date" />
+        <Inputs />
     </CreateTemplate>
 );
 
-export default withStyles(styles)(VaccinationsCreate);
+export default VaccinationsCreate;
