+++ src/version/plugins/Vaccinations/VaccinationsEdit.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { DisabledInput, TextInput, DateInput, LongTextInput } from "react-admin";
 
 import EditTemplate from "../../../core/common/ResourseTemplates/EditTemplate";
+import Inputs from "./fragments/Inputs";
 
 /**
  * This component returns block with edit form for Vaccinations
@@ -13,13 +13,7 @@ import EditTemplate from "../../../core/common/ResourseTemplates/EditTemplate";
  */
 const VaccinationsEdit = ({ classes, ...rest }) => (
     <EditTemplate blockTitle="Vaccination"  {...rest}>
-        <TextInput source="vaccinationName" label="Name" />
-        <DateInput source="vaccinationDateTime" label="Date and Time" />
-        <TextInput source="series" label="Series" />
-        <LongTextInput source="comment" label="Comment" />
-        <DisabledInput source="author" label="Author" />
-        <DisabledInput source="source" label="Source" />
-        <DisabledInput source="date" label="Date" />
+        <Inputs />
     </EditTemplate>
 );
